fix linter warning w/ static cast

+++ src/UFHeuristic.cpp
@@ -40,7 +40,7 @@ void UFHeuristic::decode(const gf2Vec& syndrome) {
     if (syndrome.size() > this->getCode()->gethZ()->pcm->size()) {
         std::vector<bool> xSyndr;
         std::vector<bool> zSyndr;
-        auto              mid = syndrome.begin() + (std::size(syndrome) / 2U);
+        auto              mid = syndrome.begin() + (static_cast<std::int64_t>(std::size(syndrome)) / 2U);
         std::move(syndrome.begin(), mid, std::back_inserter(xSyndr));
         std::move(mid, syndrome.end(), std::back_inserter(zSyndr));
         doDecoding(xSyndr, this->getCode()->gethZ());
